pyproject.toml
[project]
name = "auto-gushici"
version = "0.1.0"
description = "古诗词短视频自动剪辑系统"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "autogen-agentchat>=0.7.5",
    "autogen-ext[azure,openai]>=0.7.5",
    "dashscope>=1.25.2",
    "gradio>=6.0.1",
    "moviepy>=2.2.1",
    "pyjianyingdraft>=0.2.5",
    "python-dotenv>=1.0.1",
]
